data-governance-app/requirements.txt
streamlit>=1.28.0
snowflake-connector-python>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0
loguru>=0.7.0
cryptography>=41.0.0
plotly>=5.0.0
altair>=5.0.0
reportlab>=4.0.0
graphviz>=0.20.0
openpyxl>=3.1.2
authlib>=1.2.0
requests>=2.31.0
snowflake-snowpark-python>=1.15.0